Improve test code
This commit provides structural improvement to test code.

MOVED: Image creation functions to separate file [ImageHelper.swift].

IMPROVED: Test code using to test [DarkModeImageView].
